moved dataset.py and sampler.py to utils folder, changed modules in main.py to reflect new module hierarchy
